// typescript/interfaces.ts
export type ServiceListInterface = {
  success: boolean
  message: string
  data: Array<{
    _id: string
    user_id: string
    vendor_data: {
      _id: string
      uid: string
      fullName: string
      firstName: string
      lastName: string
      profilePicture: string
      type: string
      country: string
      replyTime: string
      languages: Array<any>
      createdAt: string
      lastDelivery: any
      lastOnline: string
    }
    package_pricing: Array<number>
    price_low: number
    price_high: number
    title: string
    description: string
    category: string
    sub_category: string
    images: Array<string>
    rating_data: {
      total_count: number
      avg_rating: number
    }
    rating: number
    address: string
    country: string
    city: string
    state: string
    zip_code: string
    excluded_areas: Array<any>
    operating_area: string
    availability: {
      monday: {
        is_available: boolean
        time_slots: Array<{
          start_time: string
          end_time: string
          start_hour: number
          start_minute: number
          end_hour: number
          end_minute: number
          max_reservation: number
          _id: string
        }>
        _id: string
      }
      tuesday: {
        is_available: boolean
        time_slots: Array<{
          start_time: string
          end_time: string
          start_hour: number
          start_minute: number
          end_hour: number
          end_minute: number
          max_reservation: number
          _id: string
        }>
        _id: string
      }
      wednesday: {
        is_available: boolean
        time_slots: Array<{
          start_time: string
          start_hour: number
          start_minute: number
          end_time: string
          end_hour: number
          end_minute: number
          max_reservation: number
          _id: string
        }>
        _id: string
      }
      thursday: {
        is_available: boolean
        time_slots: Array<{
          start_time: string
          end_time: string
          start_hour: number
          start_minute: number
          end_hour: number
          end_minute: number
          max_reservation: number
          _id: string
        }>
        _id: string
      }
      friday: {
        is_available: boolean
        time_slots: Array<{
          start_time: string
          end_time: string
          start_hour: number
          start_minute: number
          end_hour: number
          end_minute: number
          max_reservation: number
          _id: string
        }>
        _id: string
      }
      saturday: {
        is_available: boolean
        time_slots: Array<any>
        _id: string
      }
      sunday: {
        is_available: boolean
        time_slots: Array<{
          start_time: string
          start_hour: number
          start_minute: number
          end_time: string
          end_hour: number
          end_minute: number
          max_reservation: number
          _id: string
        }>
        _id?: string
      }
      _id?: string
    }
    booking_type: string
    current_availability: any
    currency: string
    createdAt: string
    filterDate: string
    updatedAt: string
    status: string
    isDeleted: boolean
    cover_image?: string
  }>
  limit: number
  page: number
  pages: number
  total: number
  grouped_counts: {
    all: number
    company: number
    individual: number
  }
}